refactor(wallet): satisfy fallow gates and rebase onto staging
Extract send/mint form hooks, shared balance invalidation helpers, and
move-usdc API so the wallet receive/send branch passes pre-push health checks.
