+++ src/features/wallet/index.ts
@@ -1,3 +1,2 @@
 export { ReceiveFundsModal } from './components/receive-funds-modal'
 export { SendTokenModal } from './components/send-token-modal'
-export { buildErc20TransferOp } from './api/build-erc20-transfer-op'
